Moved harmony patch to its own file
