+++ Beat Saber Utils/Gameplay/HarmonyPatches/StandardLevelScenesTransitionSetupDataSOFinish.cs
@@ -1,0 +1,19 @@
+﻿using Harmony;
+using System;
+
+namespace BS_Utils.Gameplay.HarmonyPatches
+{
+    [HarmonyPatch(typeof(StandardLevelScenesTransitionSetupDataSO))]
+    [HarmonyPatch("Finish", MethodType.Normal)]
+    class StandardLevelScenesTransitionSetupDataSOFinishPatch
+    {
+        static void Prefix(LevelCompletionResults levelCompletionResults)
+        {
+            if (ScoreSubmission.disabled || ScoreSubmission.prolongedDisable)
+            {
+                //Utilities.Logger.Log("Score Submission Disabled");
+                ScoreSubmission.DisableScoreSaberScoreSubmission();
+            }
+        }
+    }
+}
